fixed synchronization bugs in ConcurrantSymbolTable, MuParamVector

+++ src/java/main/edu/cmu/ml/proppr/util/TimestampedWeight.java
@@ -5,10 +5,10 @@ public class TimestampedWeight {
 //		this.k = -1;
 //		this.wt = 0.0;
 //	}
-	public TimestampedWeight(Double value,int k) {
+	public TimestampedWeight(Double value,long k) {
 		this.wt = value;
 		this.k = k;
 	}
-	public int k;
+	public long k;
 	public double wt;
 }
